+++ packages/zowe-explorer-api/CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to the "zowe-explorer-api" extension will be documented in t
 ### New features and enhancements
 
 - Refactored UI/UX methods into standalone `Gui` module for usability and maintainability. [#1967](https://github.com/zowe/vscode-extension-for-zowe/issues/1967)
+- New API call `dataSetsMatchingPattern` to allow filtering datasets via a pattern.
 - Added `fileUtils` function to USS API to facilitate copying, moving and other file operations.
 
 ### Bug fixes